requirements.txt
matplotlib==3.10.3
pandas==2.3.0
PyGithub==2.6.1
PyGithub==2.6.1
scikit_learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
statsmodels==0.14.4
